refactor: SOSFlag 리디자인
- FeedDetailContent에서 사용하던 Chip을 StepChip으로 변경
